Fix playback of ongoing recordings
We cannot determine anymore if playback has started (and ffmpeg scanning
has finished) so we unconditionally reopen the recording if it's in progress
and removed the blocking/sleep code. This also means the user can now seek to
the end and the player will stop instead of blocking/sleeping. However this is
still better than doing no reopen at all.

Fixes #11