Work around a bug in the PVR manager
If the PVR returns any value other than `PVR_ERROR_NO_ERROR` in any of
the following functions:
  `GetCapabilities`, `GetBackendName`, `GetConnectionString`,
  `GetBackendVersion`, `GetBackendHostname` or `GetTimerTypes`
the PVR will be marked as "not ready to use". This flag *never* changes
although we notify the PVR manager about our connection to the backend.

As a workaround we simply return hardcoded values until connection has
been established.

See #91